% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussquadr.R
\docType{package}
\name{gaussquadr}
\alias{gaussquadr}
\title{gaussquadr: Gaussian Quadrature}
\description{
The \code{gaussquadr} package provides routines for multivariate Gaussian
quadrature.
}
\section{\code{gaussquadr} class}{

The main object is \code{GaussQuad}, with linear transformation methods
\code{shift}, \code{rotate}, and \code{scale} and the main quadrature method
\code{apply}.
}

