% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussquadr.R
\name{lse}
\alias{lse}
\title{Efficiently and safely computes the aggregate log-sum-exp of the vector \code{x}.}
\usage{
lse(x, na_rm = FALSE)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
\code{log(sum(exp(x)))}.
}
\description{
Efficiently and safely computes the aggregate log-sum-exp of the vector \code{x}.
}
