% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussquadr.R
\name{symm_eigen}
\alias{symm_eigen}
\title{Compute robust eigendecomposition of symmetric PD matrix \code{A}.}
\usage{
symm_eigen(A, rank_tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{rank_tol}{Tolerance for rank determination: only eigenvalues > plus(\code{rank_tol} * max(eigenvalues)) are kept. Defaults to sqrt of machine precision.}

\item{Symmetric}{positive-definite matrix to be decomposed}
}
\value{
\code{eigen} object with extra field \code{valid} indicating valid eigenvalues
}
\description{
Compute robust eigendecomposition of symmetric PD matrix \code{A}.
}
