% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussquadr.R
\name{GaussianIntegrator}
\alias{GaussianIntegrator}
\title{GaussianIntegrator: multivariate Gaussian quadrature}
\usage{
GaussianIntegrator(...)
}
\arguments{
\item{...}{Remaining parameters passed to \code{GaussQuad} initializer.}

\item{n}{Number of nodes in one dimension or \code{GaussQuad} object of
type "gaussian" to be copied.}

\item{mean}{Mean of Gaussian distribution.}

\item{variance}{Variance of Gaussian distribution. Can be an
eigendecomposition, of class \code{eigen}.}
}
\value{
A new \code{GaussianIntegrator} object.
}
\description{
Create a new \code{GaussianIntegrator} object.

Prints object information.
}
\details{
R6 class for multivariate Gaussian expectation.
}
\section{Fields}{

\describe{
\item{\code{mean}}{Mean of Gaussian distribution}

\item{\code{variance}}{Variance of Gaussian distribution}
}}

