% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaussquadr.R
\name{GaussQuad}
\alias{GaussQuad}
\title{GaussQuad: multivariate Gauss quadrature}
\usage{
GaussQuad(...)
}
\arguments{
\item{type}{Type of Gauss quadrature.}

\item{n}{Number of nodes in one dimension.}

\item{d}{Number of dimensions.}

\item{alpha}{Parameter for Jacobi, Laguerre, beta and gamma quadratures.}

\item{beta}{Parameter for Jacobi and beta quadratures.}

\item{endpts}{Endpoints to be fixed in quadrature.}

\item{prune_coef}{Coefficient controling pruning: only weights above
\code{exp(d * ((1 - pc) * log(min(w)) + pc * log(max(w))))} are kept.
Defaults to \code{1 - 1 / d}. Use \code{prune_coef = 0} to keep all
weights.}

\item{prune_eps}{Prune weights below \code{eps(sum(weights))}? Defaults to TRUE.}

\item{short}{Suppress type? Defaults to FALSE.}

\item{scale}{Scale to apply to each dimension of the grid.}

\item{rotation}{Rotation to apply to each node in the grid.}

\item{shift}{Shift to apply to each dimension of the grid.}

\item{f}{Integrand function.}
}
\value{
A new \code{GaussQuad} object.

The computed quadrature.
}
\description{
Create a new \code{GaussQuad} object.

Prints object information.

Scale the node grid of a \code{GaussQuad} object.

Rotates the node grid of a \code{GaussQuad} object.

Shift the node grid of a \code{GaussQuad} object.

Computes the quadrature of the integrand \code{f}.
}
\details{
R6 class with methods for linear transformations and integration.
}
\section{Fields}{

\describe{
\item{\code{type}}{Type of Gauss quadrature}

\item{\code{nodes}}{Grid of nodes}

\item{\code{weights}}{Quadrature weights}

\item{\code{alpha}}{Parameter for Jacobi, Laguerre, beta and gamma quadratures}

\item{\code{beta}}{Parameter for Jacobi and beta quadratures}
}}

